requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.2
aspy.refactor-imports==3.0.1
async-timeout==4.0.2
attrs==21.4.0
autoflake==1.4
base58==2.1.1
bitarray==1.2.2
black==22.3.0
certifi==2022.5.18.1
cfgv==3.3.1
charset-normalizer==2.0.12
click==8.1.3
croniter==1.3.5
cytoolz==0.11.2
distlib==0.3.4
Django==4.0.5
django-debug-toolbar==3.4.0
eth-abi==2.1.1
eth-account==0.5.7
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
filelock==3.7.1
flake8==4.0.1
frozenlist==1.3.0
hexbytes==0.2.2
identify==2.5.1
idna==3.3
ipfshttpclient==0.8.0a2
jsonschema==4.6.0
lru-dict==1.1.7
mccabe==0.6.1
multiaddr==0.0.9
multidict==6.0.2
mypy-extensions==0.4.3
netaddr==0.8.0
nodeenv==1.6.0
parsimonious==0.8.1
pathspec==0.9.0
platformdirs==2.5.2
pre-commit==2.19.0
protobuf==3.20.1
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pycryptodome==3.14.1
pyflakes==2.4.0
pymongo==4.1.1
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pyupgrade==2.32.1
PyYAML==6.0
reorder-python-imports==3.1.0
requests==2.27.1
rlp==2.0.1
shortuuid==1.0.9
six==1.16.0
sqlparse==0.4.2
tokenize-rt==4.2.1
toml==0.10.2
tomli==2.0.1
toolz==0.11.2
urllib3==1.26.9
uWSGI==2.0.20
varint==1.0.2
virtualenv==20.14.1
web3==5.29.2
websockets==9.1
yarl==1.7.2
yesqa==1.3.0
